Add ParkingLot and Runner classes for system design task
Added new files ParkingLot.java and Runner.java as part of a system design task. ParkingLot.java is designed to manage a parking lot with a given capacity and the ability to park/unpark different types of vehicles. Runner.java serves as the main class to interact with and control the ParkingLot behavior. It allows users to execute commands like PARK, UNPARK, CAPACITY and STATUS to manage parking lot functionalities.