Refactored digital wallet system design
The digital wallet system design was refactored. This involved the deletion of the 'DigitalWallet' class, creation of new classes like 'WalletDAO' and 'WalletService', and modification of the 'WalletAccount' class. The changes aim to segregate responsibilities between entities, data access objects (DAOs), and services, for more optimal organization and functionality of the code.